1. correct data type 2. use release version jar

+++ ad-data-exchanger-client/src/main/java/com/hotstar/adtech/blaze/exchanger/api/entity/AllocationPlanDetail.java
@@ -12,7 +12,7 @@ public class AllocationPlanDetail {
   PlanType planType;
   Integer totalBreakNumber;
   Integer nextBreakIndex;
-  List<Long> breakTypeIds;
+  List<Integer> breakTypeIds;
   Integer duration;
   String fileName;
   String md5;
